requirements.txt
# requirements.txt - Complete Version

# ==================== CORE ====================
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
python-dotenv==1.0.0

# ==================== DATABASES ====================
# PostgreSQL (Supabase)
psycopg2-binary==2.9.9

# Neo4j Graph Database
neo4j==5.16.0

# ==================== OPTIONAL (Future Features) ====================
# File Uploads & Forms
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Requests (if calling external APIs)
requests==2.31.0
httpx==0.26.0

# Document Generation (Step 6)
python-docx==1.1.0
fpdf2>=2.8.0
reportlab==4.0.9

# E-Signature Integration (Step 6)
docusign-esign==3.24.0

# Date/Time Utilities
python-dateutil==2.8.2

# ==================== DEVELOPMENT ====================
# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
httpx==0.26.0  # For TestClient

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
